实装：（能源）初步实装了能源管理系统的供电策略调整功能

+++ Script/UI/Panel/manage_assembly_line_panel.py
@@ -467,11 +467,11 @@ class Manage_Assembly_Line_Panel:
                 line_feed.draw()
 
             line_feed.draw()
-            yes_draw = draw.CenterButton(_("[确定]"), _("确定"), window_width / 2)
+            yes_draw = draw.CenterButton(_("[确定]"), _("确定"), int(window_width / 2))
             if self.now_chara_id != -1:
                 yes_draw.draw()
                 return_list.append(yes_draw.return_text)
-            back_draw = draw.CenterButton(_("[返回]"), _("返回"), window_width / 2)
+            back_draw = draw.CenterButton(_("[返回]"), _("返回"), int(window_width / 2))
             back_draw.draw()
             line_feed.draw()
             return_list.append(back_draw.return_text)
